ADD TASK AND REMOVE TASK IMPLEMENTED
