+++ src/modules/addItem.js
@@ -1,46 +1,5 @@
 import Delete from '../assets/delete.svg';
 
-const removeTask = () => {
-  const toRemove = document.getElementById('delButton');
-  toRemove.parentElement.remove();
-
-  localStorage.clear();
-
-  const available = document.querySelectorAll('.desc');
-  if (available.length > 0) {
-    const newArray = [];
-
-    let length = 1;
-
-    [...available].forEach((child) => {
-      const toAdd = { index: length, completed: false, desc: child.innerText };
-      newArray.push(toAdd);
-      length += 1;
-    });
-    localStorage.setItem('ITEMS', JSON.stringify(newArray));
-  }
-};
-
-const updateTask = () => {
-  //
-  localStorage.getItem('ITEMS');
-  const available = document.querySelectorAll('#checkbox');
-  if (available.length > 0) {
-    const newArray = [];
-    let length = 1;
-    [...available].forEach((child) => {
-      const toAdd = {
-        index: length,
-        completed: child.checked,
-        desc: child.nextElementSibling.innerText,
-      };
-      newArray.push(toAdd);
-      length += 1;
-    });
-    localStorage.setItem('ITEMS', JSON.stringify(newArray));
-  }
-};
-
 const AddItem = (item) => {
   const rowitem = document.createElement('div');
   rowitem.className = 'row-item';
@@ -52,13 +11,11 @@ const AddItem = (item) => {
   if (boolean.checkd) {
     boolean.click();
   }
-  boolean.addEventListener('change', updateTask);
 
   const desc = document.createElement('div');
   desc.className = 'desc';
   desc.innerText = item.desc;
   desc.contentEditable = true;
-  desc.addEventListener('change', updateTask);
 
   const delImg = document.createElement('img');
   delImg.setAttribute('src', Delete);
@@ -69,7 +26,6 @@ const AddItem = (item) => {
   const delButton = document.createElement('button');
   delButton.setAttribute('id', 'delButton');
   delButton.appendChild(delImg);
-  delButton.addEventListener('click', removeTask);
 
   rowitem.appendChild(boolean);
   rowitem.appendChild(desc);
